requirements.txt
Django==5.0
pre-commit==3.5.0
cfgv==3.4.0
identify==2.5.32
nodeenv==1.8.0
platformdirs==4.1.0
filelock==3.13.1
PyYAML==6.0.1
virtualenv==20.25.0
distlib==0.3.7
pytz==2023.3.post1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
sqlparse==0.4.4
typing_extensions==4.8.0
python-dotenv==1.0.0
drf-spectacular==0.26.5
uritemplate==4.1.1
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
attrs==23.1.0
importlib-resources==6.1.1
rpds-py==0.13.2
zipp==3.17.0
pkgutil_resolve_name==1.3.10
referencing==0.31.1
psycopg2-binary==2.9.9
django-timescaledb==0.2.13
factory-boy==3.3.0